nits

+++ include/afl-fuzz.h
@@ -1405,7 +1405,7 @@ void   check_if_tty(afl_state_t *);
 void   save_cmdline(afl_state_t *, u32, char **);
 void   read_foreign_testcases(afl_state_t *, int);
 void   write_crash_readme(afl_state_t *afl);
-u8     check_if_text_buf(u8 *buf, u32 len);
+u32    check_if_text_buf(u8 *buf, u32 len);
 #ifndef AFL_SHOWMAP
 void setup_signal_handlers(void);
 #endif
